fix: Query Editor toolbar folds to menu to fit the window size
