+++ src/webviews/QueryEditor/QueryEditor.tsx
@@ -17,7 +17,7 @@ const useStyles = makeStyles({
     root: {
         display: 'grid',
         gridTemplateRows: '100vh',
-        // minWidth: '900px',
+        // minWidth: '520px',
     },
 });
