+++ src/webviews/Timer.tsx
@@ -4,7 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 export type TimerProps = {
-    time: number;
+    time: number; // time in milliseconds
 };
 
 export const Timer = (props: TimerProps) => {
@@ -12,7 +12,7 @@ export const Timer = (props: TimerProps) => {
         <div className="timer">
             <span className="digits">{('0' + Math.floor((props.time / 60000) % 60)).slice(-2)}:</span>
             <span className="digits">{('0' + Math.floor((props.time / 1000) % 60)).slice(-2)}.</span>
-            <span className="digits mili-sec">{('0' + ((props.time / 10) % 100)).slice(-2)}</span>
+            <span className="digits mili-sec">{('0' + Math.floor((props.time / 10) % 100)).slice(-2)}</span>
         </div>
     );
 };
